Changed the URL handling to follow the specification from GRC.com. The specification states that domain is the path without sqrl:// before either the slash or pipe sign. And we should not care about either port or username and password in this string.
And when it comes to the url we call back to we need to send the
querystring so servers could retrieve data from the querystring that might
be needed for processing.

Also added the flag to permit all in strict mode to be able to run this
before we do the web calles in a background process.